refactor: store-reducer 분리
