+++ __mock__/react-redux.js
@@ -1,3 +1,3 @@
 export const useDispatch = jest.fn();
 
-export const useSelector = jest.fn();
+export const useSelector = jest.fn((selector) => selector({}));
